Implemented AddAssignment component
